Wrappeo de create y delete de Stk::FileWvOut. Arreglo con tipos.

+++ EsferixisMusicFrameworkBackend/EMFB_STK/include/EMFB_STK_FILEWRITE.h
@@ -2,4 +2,4 @@
 
 #include "EMFB_STK_COMMON.h"
 
-EMFB_STK_API int emfb_stk_filewrite_FILE_WAV();
+EMFB_STK_API long emfb_stk_filewrite_FILE_WAV();
